Update to iOs 7.1. Get Crashlytics from Pod. Rename to fix warnings.

+++ KangxiRadicals/Kangxi Radicals/Categories/UIViewController+Additions.h
@@ -9,5 +9,5 @@
 #import <UIKit/UIKit.h>
 
 @interface UIViewController (Additions)
--(UIView *)titleViewWithText:(NSString *)title numberOfChineseCharacters:(int)n;
+-(UIView *)titleViewWithText:(NSString *)title numberOfChineseCharacters:(NSInteger)n;
 @end
